app/requirements.txt
fastapi==0.111.0
pandas==2.2.2
psycopg2-binary==2.9.9 
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
sqlmodel==0.0.19
pytest==8.2.2 
